Rewritten ld-comb-pal PAL colour filter for full 16-bit input TBC support

+++ tools/ld-comb-pal/palcolour.h
@@ -62,7 +62,7 @@ private:
     double yfilt[4][arraySize + 1];
     double cdiv;
     double ydiv;
-    int32_t refAmpl;
+    double refAmpl;
 
     // Method to build the required look-up tables
     void buildLookUpTables(void);
